+++ RELEASE_NOTES.md
@@ -2,3 +2,5 @@ v0.88.1
 
 - backups: Add hasSeenAdminDeleteEducationDialog field support.
 
+- java: Classify `ChatServiceInactive` errors as `RequestResult.RetryableNetworkError`
+